+++ benchopt/parallel_backends/slurm_executor.py
@@ -92,14 +92,19 @@ def run_on_slurm(
     with ExitStack() as stack:
         for kwargs in all_runs:
             solver = kwargs.get("solver")
-            solver_slurm_config = get_solver_slurm_config(solver, slurm_config)
-            executor_config = hashable_pytree(solver_slurm_config)
+            if solver is not None:
+                job_slurm_config = get_solver_slurm_config(
+                    solver, slurm_config
+                )
+            else:
+                job_slurm_config = slurm_config
+            executor_config = hashable_pytree(job_slurm_config)
 
             if executor_config not in executors:
                 executor = get_slurm_executor(
                     benchmark,
-                    solver_slurm_config,
-                    timeout=common_kwargs["timeout"],
+                    job_slurm_config,
+                    timeout=common_kwargs.get("timeout"),
                 )
                 stack.enter_context(executor.batch())
                 executors[executor_config] = executor